# Sito Web/api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3100",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/register": {
      "put": {
        "tags": [
          "Autentificazione"
        ],
        "summary": "Registra l'utente dentro",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Autentificazione"
        ],
        "summary": "Fase di controllo per il login",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/add_favorite_artist": {
      "put": {
        "tags": [
          "Artista"
        ],
        "summary": "Aggiunge l'artista ai preferiti per quell'account [Autentificazione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "artist": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/view_favorite_artist": {
      "post": {
        "tags": [
          "Artista"
        ],
        "summary": "Guarda gli artisti preferiti per quell'account [Autentificazione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/delete_favorite_artist": {
      "delete": {
        "tags": [
          "Artista"
        ],
        "summary": "Rimuove l'artista ai preferiti per quell'account [Autentificazione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "artist": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/search_genre": {
      "post": {
        "tags": [
          "Cerca"
        ],
        "summary": "Cerca in base ai generi di spotify con quello dato",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "genres": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/add_genre": {
      "put": {
        "tags": [
          "Genere"
        ],
        "summary": "Aggiunge generi ai preferiti [Autentificazione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "genre": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/view_favorite_genre": {
      "post": {
        "tags": [
          "Genere"
        ],
        "summary": "Visualizza i genere preferiti [Autentificazione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/delete_favorite_genre": {
      "delete": {
        "tags": [
          "Genere"
        ],
        "summary": "Cancella il genere preferito [Autentificazione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "genre": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/account_info": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Restituisce le informazioni dell'account [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/remove_account": {
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Rimuove l'account [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/change_profilo": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Cambia i dati inerenti all'account [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/view_suggestion": {
      "post": {
        "tags": [
          "Home"
        ],
        "summary": "Restituisce i brani correlati ai genere e artisti preferiti [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/view_my_playlist": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "summary": "Restituisce le informazioni delle playlist che segue / ha creato l'account [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "creator": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/delete_tag": {
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Rimuove i tag dalla playlist interna [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "example": "any"
                },
                "tag": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/add_tag": {
      "put": {
        "tags": [
          "Tag"
        ],
        "summary": "Aggiunge i tag alla playlist interna [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "example": "any"
                },
                "tag": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/modifica_playlist": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "summary": "Modifico la playlist interna [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "example": "any"
                },
                "creator": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "desc": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/delete_playlist": {
      "delete": {
        "tags": [
          "Playlist"
        ],
        "summary": "Rimuovo la playlist interna [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "example": "any"
                },
                "creator": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/modifica_privato": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "summary": "Modifico la playlist interna e la rendo o privata o pubblica [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "creator": {
                  "example": "any"
                },
                "public": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/remove_track": {
      "delete": {
        "tags": [
          "Playlist"
        ],
        "summary": "Rimuove il brano dalla playlist interna [Autentificzione richiesta]",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "playlist": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
  }
}